+++ NEWS.md
@@ -1,4 +1,5 @@
 # MetricGraph (development version)
+* Changed argument `data` to `newdata` in `predict` methods. The argument `data` was deprecated.
 
 # MetricGraph 1.1.2
 * Adjusts to ensure compatibility with future releases of the Matrix package.